+++ tests/Cache/RouteMetadata/Fixtures/DummyClass.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Sofascore\PurgatoryBundle\Tests\Cache\RouteMetadata\Fixtures;
+
+class DummyClass
+{
+    public static function getValues()
+    {
+    }
+}